feat(allure-js-commons): move 'serialize' to /sdk/utils
This will allow using 'serialize' from environments other than Node.js.
